#140
- удален SelectScreen и SelectScreenProps (нужно использовать SelectScreenLayout и SelectScreenOption)
- правки stories
